+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 ## v1.0.3
 - `Configuration.getProperty()` must provide 3 parameters.
+- Add `"Unsupported or Error"` response on sensors.
 
 ## v1.0.2
 - Add `global` wrapping for namespace and classes to prevent bundlers trying to import.